% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiff.R
\name{PlotDiff}
\alias{PlotDiff}
\title{Plot differences across mixture models}
\usage{
PlotDiff(object, type = "weight", dim = c(1, 2),
  group = 1:length(unique(object$data$C)), xlim = range(object$data$Y[,
  dim[1]]), ylim = range(object$data$Y[, dim[2]]), ...)
}
\arguments{
\item{object}{Object of class inheriting from "cremid".}
}
\value{
A plot.
}
\description{
Plot differences across mixture models
}
\examples{

n = c(250, 250)
p = 4

Y1 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 3, ncol = p))
Y2 = rbind( matrix( rnorm( n[1]*p), ncol = p), matrix( rnorm(n[2]*p) + 4, ncol = p))
Y = rbind(Y1, Y2)
C = c( rep(1,sum(n)), rep(2,sum(n)))

ans = Fit(Y, C)
PlotDiff(ans, type = "weight")
PlotDiff(ans, type = "shift")
}
